Structure Chernobog vehicles better
