+++ addons/chernobog/CfgVehicles_Vehicles_Children.hpp
@@ -1,0 +1,92 @@
+// Offroad
+class TACU_Chernobog_Children_V_I_Offroad: TACU_Offroad_Unarmed_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    textureList[] = {
+        "Guerilla_01", 0.14,
+        "Guerilla_04", 0.14,
+        "Guerilla_05", 0.14,
+        "Guerilla_07", 0.14,
+        "Guerilla_08", 0.14,
+        "Guerilla_09", 0.14,
+        "Guerilla_12", 0.14
+    };
+    EDITORPREVIEW(TACU_Chernobog_Children_V_I_Offroad);
+};
+// HMG Offroad
+class TACU_Chernobog_Children_V_I_Offroad_HMG: TACU_Offroad_Armed_M2_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+    textureList[] = {
+        "Guerilla_01", 0.14,
+        "Guerilla_04", 0.14,
+        "Guerilla_05", 0.14,
+        "Guerilla_07", 0.14,
+        "Guerilla_08", 0.14,
+        "Guerilla_09", 0.14,
+        "Guerilla_12", 0.14
+    };
+    EDITORPREVIEW(TACU_Chernobog_Children_V_I_Offroad_HMG);
+};
+class TACU_Chernobog_Children_V_I_UAZ: TACU_UAZ_Open_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    //EDITORPREVIEW(TACU_Chernobog_Children_V_I_UAZ);
+};
+class TACU_Chernobog_Children_V_I_UAZ_DSHKM: TACU_UAZ_DSHKM_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    //EDITORPREVIEW(TACU_Chernobog_Children_V_I_UAZ_DSHKM);
+};
+class TACU_Chernobog_Children_V_I_Skoda: TACU_S1203_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    //EDITORPREVIEW(TACU_Chernobog_Children_V_I_Skoda);
+};
+class TACU_Chernobog_Children_V_I_Ural: TACU_Ural_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    //EDITORPREVIEW(TACU_Chernobog_Children_V_I_Ural);
+};
+class TACU_Chernobog_Children_V_I_Hilux: TACU_Hilux_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    //EDITORPREVIEW(TACU_Chernobog_Children_V_I_Hilux);
+};
+class TACU_Chernobog_Children_V_I_Hilux_Armored: TACU_Hilux_Armored_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Children_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Children_U_I_Rifleman_01";
+    typicalCargo[] = {"TACU_Chernobog_Children_U_I_Rifleman_01"};
+    //EDITORPREVIEW(TACU_Chernobog_Children_V_I_Hilux_Armored);
+};
