+++ addons/chernobog/CfgVehicles_Vehicles_Chosen.hpp
@@ -1,0 +1,86 @@
+class TACU_Chernobog_Chosen_V_I_Offroad: TACU_Chernobog_Children_V_I_Offroad {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_Offroad_HMG: TACU_Chernobog_Children_V_I_Offroad_HMG {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+// UP Armoured Offroad
+class TACU_Chernobog_Chosen_V_I_Offroad_UP: TACU_Offroad_Unarmed_UP_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+    textureList[] = {
+        "Guerilla_01", 0.14,
+        "Guerilla_04", 0.14,
+        "Guerilla_05", 0.14,
+        "Guerilla_07", 0.14,
+        "Guerilla_08", 0.14,
+        "Guerilla_09", 0.14,
+        "Guerilla_12", 0.14
+    };
+    EDITORPREVIEW(TACU_Chernobog_Chosen_V_I_Offroad_UP);
+};
+
+
+// UP Armoured Offroad HMG
+class TACU_Chernobog_Chosen_V_I_Offroad_HMG_UP: TACU_Offroad_Armed_M2_UP_Base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Chernobog";
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    side = 2;
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+    textureList[] = {
+        "Guerilla_01", 0.14,
+        "Guerilla_04", 0.14,
+        "Guerilla_05", 0.14,
+        "Guerilla_07", 0.14,
+        "Guerilla_08", 0.14,
+        "Guerilla_09", 0.14,
+        "Guerilla_12", 0.14
+    };
+    EDITORPREVIEW(TACU_Chernobog_Chosen_V_I_Offroad_HMG_UP);
+};
+class TACU_Chernobog_Chosen_V_I_UAZ: TACU_Chernobog_Children_V_I_UAZ {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_UAZ_DSHKM: TACU_Chernobog_Children_V_I_UAZ_DSHKM {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_Skoda: TACU_Chernobog_Children_V_I_Skoda {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_Ural: TACU_Chernobog_Children_V_I_Ural {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_Hilux: TACU_Chernobog_Children_V_I_Hilux {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_Hilux_Armored: TACU_Chernobog_Children_V_I_Hilux_Armored {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+};
+class TACU_Chernobog_Chosen_V_I_Hilux_BMP1: TACU_Hilux_BMP1_Base {
+    editorSubcategory = "TACU_Chernobog_Chosen_Cars";
+    crew = "TACU_Chernobog_Chosen_U_I_Rifleman";
+    typicalCargo[] = {"TACU_Chernobog_Chosen_U_I_Rifleman"};
+    //EDITORPREVIEW(TACU_Chernobog_Chosen_V_I_Hilux_BMP1);
+};
